Add files via upload
In this script, you can connect to your smart home hub using its IP address and then define functions to control your smart home devices. The TurnOnLight and TurnOffLight functions take a room name as a parameter and send a command to the smart home hub to turn on or off the light in that room. The SetThermostat function takes a temperature as a parameter and sends a command to the smart home hub to set the thermostat to that temperature.

After defining the functions, the script prompts the user for input and then executes the user's choice by calling the appropriate function. Finally, the script closes the connection to the smart home hub.